+++ packages/react/runtime/src/lifecycle/patch/commit.ts
@@ -21,17 +21,23 @@
 
 import { options } from 'preact';
 
+import type { RunWorkletCtxData } from '@lynx-js/react/worklet-runtime/bindings';
+
 import { LifecycleConstant } from '../../lifecycleConstant.js';
 import { globalPipelineOptions, markTiming, markTimingLegacy, setPipeline } from '../../lynx/performance.js';
 import { COMMIT } from '../../renderToOpcodes/constants.js';
 import { applyQueuedRefs } from '../../snapshot/ref.js';
 import { backgroundSnapshotInstanceManager } from '../../snapshot.js';
 import { hook, isEmptyObject } from '../../utils.js';
-import { takeWorkletRefInitValuePatch } from '../../worklet/workletRefPool.js';
+import { takeWorkletRefInitValuePatch } from '../../worklet/ref/workletRefPool.js';
 import { getReloadVersion } from '../pass.js';
 import type { SnapshotPatch } from './snapshotPatch.js';
 import { takeGlobalSnapshotPatch } from './snapshotPatch.js';
 import { profileEnd, profileStart } from '../../debug/utils.js';
+import {
+  delayedRunOnMainThreadData,
+  takeDelayedRunOnMainThreadData,
+} from '../../worklet/call/delayedRunOnMainThreadData.js';
 import { isRendering } from '../isRendering.js';
 
 let globalFlushOptions: FlushOptions = {};
@@ -62,6 +68,7 @@ interface Patch {
  */
 interface PatchList {
   patchList: Patch[];
+  delayedRunOnMainThreadData?: RunWorkletCtxData[];
   flushOptions?: FlushOptions;
 }
 
@@ -155,6 +162,9 @@ function replaceCommitHook(): void {
       };
       if (!isEmptyObject(flushOptions)) {
         patchList.flushOptions = flushOptions;
+      }
+      if (snapshotPatch && delayedRunOnMainThreadData.length) {
+        patchList.delayedRunOnMainThreadData = takeDelayedRunOnMainThreadData();
       }
       const obj = commitPatchUpdate(patchList, patchOptions);