+++ packages/react/runtime/src/list.ts
@@ -153,6 +153,11 @@ export function componentAtIndexFactory(
         }]);
       }
       const root = childCtx.__element_root!;
+      // In the defer `list-item` scenario, `componentAtIndex` occurs with delay.
+      // Within `componentAtIndex`, nodes that quickly appear and disappear due to re-layout will be enqueued again,
+      // causing the mapping relationship between sign and SnapshotInstance to become corrupted.
+      // This results in a SnapshotInstance without `__elements` being enqueued.
+      signMap.set(sign, childCtx);
       if (!enableBatchRender) {
         const flushOptions: FlushOptions = {
           triggerLayout: true,
@@ -179,7 +184,6 @@ export function componentAtIndexFactory(
         }
         __FlushElementTree(root, flushOptions);
       }
-      signMap.set(sign, childCtx);
       return sign;
     }
 
@@ -187,6 +191,7 @@ export function componentAtIndexFactory(
     const root = childCtx.__element_root!;
     __AppendElement(list, root);
     const sign = __GetElementUniqueID(root);
+    signMap.set(sign, childCtx);
     if (!enableBatchRender) {
       __FlushElementTree(root, {
         triggerLayout: true,
@@ -199,7 +204,6 @@ export function componentAtIndexFactory(
         asyncFlush: true,
       });
     }
-    signMap.set(sign, childCtx);
     return sign;
   };