+++ packages/react/runtime/src/lynx/tt.ts
@@ -21,6 +21,10 @@ import { CHILDREN } from '../renderToOpcodes/constants.js';
 import { __root } from '../root.js';
 import { backgroundSnapshotInstanceManager } from '../snapshot.js';
 import type { SerializedSnapshotInstance } from '../snapshot.js';
+import {
+  delayedRunOnMainThreadData,
+  takeDelayedRunOnMainThreadData,
+} from '../worklet/call/delayedRunOnMainThreadData.js';
 import { destroyWorklet } from '../worklet/destroy.js';
 
 export { runWithForce };
@@ -120,6 +124,9 @@ function onLifecycleEventImpl(type: LifecycleConstant, data: unknown): void {
       const patchList: PatchList = {
         patchList: [{ snapshotPatch, id: commitTaskId }],
       };
+      if (delayedRunOnMainThreadData.length) {
+        patchList.delayedRunOnMainThreadData = takeDelayedRunOnMainThreadData();
+      }
       const obj = commitPatchUpdate(patchList, { isHydration: true });
 
       lynx.getNativeApp().callLepusMethod(LifecycleConstant.patchUpdate, obj, () => {