+++ packages/rspeedy/core/src/config/rsbuild/index.ts
@@ -57,6 +57,7 @@ export function toRsbuildConfig(
     },
     resolve: {
       alias: config.resolve?.alias,
+      aliasStrategy: config.resolve?.aliasStrategy,
       dedupe: config.resolve?.dedupe,
     },
     source: {
@@ -88,6 +89,8 @@ export function toRsbuildConfig(
       host: config.server?.host,
 
       port: config.server?.port,
+
+      proxy: config.server?.proxy,
 
       strictPort: config.server?.strictPort,
     },