+++ packages/web-platform/web-constants/src/types/LynxModule.ts
@@ -38,6 +38,7 @@ export interface LynxTemplate {
   };
   elementTemplate: Record<string, ElementTemplateData[]>;
   version?: number;
+  appType: 'card' | 'lazy';
 }
 
 export interface LynxJSModule {