+++ packages/rspeedy/plugin-react-alias/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lynx-js/react-alias-rsbuild-plugin
 
+## 0.10.14
+
+### Patch Changes
+
+- Allow customization of the react$ alias. ([#1653](https://github.com/lynx-family/lynx-stack/pull/1653))
+
+  ```js
+  import { defineConfig } from '@lynx-js/rspeedy'
+
+  export default defineConfig({
+    resolve: {
+      alias: {
+        react$: '@lynx-js/react/compat',
+      },
+    },
+  })
+  ```
+
 ## 0.10.13
 
 ### Patch Changes