+++ packages/web-platform/web-style-transformer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @lynx-js/web-style-transformer
 
+## 0.16.1
+
 ## 0.16.0
 
 ## 0.15.7